+++ fixtures/js/css_import.js
@@ -1,0 +1,2 @@
+
+require('./../css/h1_style.css');