+++ test/config-generator.js
@@ -220,12 +220,6 @@ describe('The config-generator function', () => {
             config.enableVersioning(true);
 
             const actualConfig = configGenerator(config);
-
-            const loaderOptionsPlugin = findPlugin(webpack.LoaderOptionsPlugin, actualConfig.plugins);
-            expect(loaderOptionsPlugin.options.debug).to.equal(true);
-            expect(loaderOptionsPlugin.options.options.context).to.equal('/tmp/context');
-            expect(loaderOptionsPlugin.options.options.output.path).to.equal('/tmp/output/public-path');
-
             expect(actualConfig.optimization.namedModules).to.be.true;
         });
 
@@ -238,9 +232,6 @@ describe('The config-generator function', () => {
             config.addEntry('main', './main');
 
             const actualConfig = configGenerator(config);
-
-            const loaderOptionsPlugin = findPlugin(webpack.LoaderOptionsPlugin, actualConfig.plugins);
-            expect(loaderOptionsPlugin.options.debug).to.equal(false);
 
             const moduleHashedIdsPlugin = findPlugin(webpack.HashedModuleIdsPlugin, actualConfig.plugins);
             expect(moduleHashedIdsPlugin).to.not.be.undefined;
@@ -865,7 +856,7 @@ describe('The config-generator function', () => {
 
             const babelLoader = jsRule.use.find(loader => loader.loader === 'babel-loader');
             const babelEnvPreset = babelLoader.options.presets.find(([name]) => name === '@babel/preset-env');
-            expect(babelEnvPreset[1].useBuiltIns).to.equal('entry');
+            expect(babelEnvPreset[1].useBuiltIns).to.equal(false);
         });
 
         it('with configureBabel() and a different exclude rule', () => {
@@ -906,7 +897,8 @@ describe('The config-generator function', () => {
             config.publicPath = '/public-path';
             config.addEntry('main', './main');
             config.configureBabel(() => { }, {
-                useBuiltIns: 'usage'
+                useBuiltIns: 'usage',
+                corejs: 3,
             });
 
             const actualConfig = configGenerator(config);
@@ -915,6 +907,7 @@ describe('The config-generator function', () => {
             const babelLoader = jsRule.use.find(loader => loader.loader === 'babel-loader');
             const babelEnvPreset = babelLoader.options.presets.find(([name]) => name === '@babel/preset-env');
             expect(babelEnvPreset[1].useBuiltIns).to.equal('usage');
+            expect(babelEnvPreset[1].corejs).to.equal(3);
         });
     });
 
@@ -1012,6 +1005,209 @@ describe('The config-generator function', () => {
                 ignored: /node_modules/,
                 poll: 250,
             });
+        });
+    });
+
+    describe('Test configureLoaderRule()', () => {
+        let config;
+
+        beforeEach(() => {
+            config = createConfig();
+            config.outputPath = '/tmp/public/build';
+            config.setPublicPath('/');
+            config.enableSingleRuntimeChunk();
+        });
+
+        it('configure rule for "javascript"', () => {
+            config.configureLoaderRule('javascript', (loaderRule) => {
+                loaderRule.test = /\.m?js$/;
+                loaderRule.use[0].options.fooBar = 'fooBar';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.m?js$/, webpackConfig.module.rules);
+
+            expect('file.js').to.match(rule.test);
+            expect('file.mjs').to.match(rule.test);
+            expect(rule.use[0].options.fooBar).to.equal('fooBar');
+        });
+
+        it('configure rule for the alias "js"', () => {
+            config.configureLoaderRule('js', (loaderRule) => {
+                loaderRule.test = /\.m?js$/;
+                loaderRule.use[0].options.fooBar = 'fooBar';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.m?js$/, webpackConfig.module.rules);
+
+            expect('file.js').to.match(rule.test);
+            expect('file.mjs').to.match(rule.test);
+            expect(rule.use[0].options.fooBar).to.equal('fooBar');
+        });
+
+        it('configure rule for "css"', () => {
+            config.configureLoaderRule('css', (loaderRule) => {
+                loaderRule.camelCase = true;
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.css$/, webpackConfig.module.rules);
+
+            expect(rule.camelCase).to.be.true;
+        });
+
+        it('configure rule for "images"', () => {
+            config.configureLoaderRule('images', (loaderRule) => {
+                loaderRule.options.name = 'dirname-images/[hash:42].[ext]';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.(png|jpg|jpeg|gif|ico|svg|webp)$/, webpackConfig.module.rules);
+
+            expect(rule.options.name).to.equal('dirname-images/[hash:42].[ext]');
+        });
+
+        it('configure rule for "fonts"', () => {
+            config.configureLoaderRule('fonts', (loader) => {
+                loader.options.name = 'dirname-fonts/[hash:42].[ext]';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.(woff|woff2|ttf|eot|otf)$/, webpackConfig.module.rules);
+
+            expect(rule.options.name).to.equal('dirname-fonts/[hash:42].[ext]');
+        });
+
+        it('configure rule for "sass"', () => {
+            config.enableSassLoader();
+            config.configureLoaderRule('sass', (loaderRule) => {
+                loaderRule.oneOf[1].use[2].options.fooBar = 'fooBar';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.s[ac]ss$/, webpackConfig.module.rules);
+
+            expect(rule.oneOf[1].use[2].options.fooBar).to.equal('fooBar');
+        });
+
+        it('configure rule for the alias "scss"', () => {
+            config.enableSassLoader();
+            config.configureLoaderRule('scss', (loaderRule) => {
+                loaderRule.oneOf[1].use[2].options.fooBar = 'fooBar';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.s[ac]ss$/, webpackConfig.module.rules);
+
+            expect(rule.oneOf[1].use[2].options.fooBar).to.equal('fooBar');
+        });
+
+        it('configure rule for "less"', () => {
+            config.enableLessLoader((options) => {
+                options.optionA = 'optionA';
+            });
+            config.configureLoaderRule('less', (loaderRule) => {
+                loaderRule.oneOf[1].use[2].options.optionB = 'optionB';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.less/, webpackConfig.module.rules);
+
+            expect(rule.oneOf[1].use[2].options.optionA).to.equal('optionA');
+            expect(rule.oneOf[1].use[2].options.optionB).to.equal('optionB');
+        });
+
+        it('configure rule for "stylus"', () => {
+            config.enableStylusLoader((options) => {
+                options.optionA = 'optionA';
+            });
+            config.configureLoaderRule('stylus', (loaderRule) => {
+                loaderRule.oneOf[1].use[2].options.optionB = 'optionB';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.styl/, webpackConfig.module.rules);
+
+            expect(rule.oneOf[1].use[2].options.optionA).to.equal('optionA');
+            expect(rule.oneOf[1].use[2].options.optionB).to.equal('optionB');
+        });
+
+        it('configure rule for "vue"', () => {
+            config.enableVueLoader((options) => {
+                options.shadowMode = true;
+            });
+            config.configureLoaderRule('vue', (loaderRule) => {
+                loaderRule.use[0].options.prettify = false;
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.vue$/, webpackConfig.module.rules);
+
+            expect(rule.use[0].options.shadowMode).to.be.true;
+            expect(rule.use[0].options.prettify).to.be.false;
+        });
+
+        it('configure rule for "eslint"', () => {
+            config.enableEslintLoader((options) => {
+                options.extends = 'airbnb';
+            });
+            config.configureLoaderRule('eslint', (loaderRule) => {
+                loaderRule.test = /\.(jsx?|vue)/;
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.(jsx?|vue)/, webpackConfig.module.rules);
+
+            expect(rule.options.extends).to.equal('airbnb');
+            expect('file.js').to.match(rule.test);
+            expect('file.jsx').to.match(rule.test);
+            expect('file.vue').to.match(rule.test);
+        });
+
+        it('configure rule for "typescript" and "ts"', () => {
+            config.enableTypeScriptLoader((options) => {
+                options.silent = true;
+            });
+            config.configureLoaderRule('typescript', (loaderRule) => {
+                loaderRule.use[1].options.happyPackMode = true;
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.tsx?$/, webpackConfig.module.rules);
+
+            expect(rule.use[1].options.silent).to.be.true;
+            expect(rule.use[1].options.happyPackMode).to.be.true;
+        });
+
+        it('configure rule for the alias "ts"', () => {
+            config.enableTypeScriptLoader((options) => {
+                options.silent = true;
+            });
+            config.configureLoaderRule('ts', (loaderRule) => {
+                loaderRule.use[1].options.happyPackMode = true;
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.tsx?$/, webpackConfig.module.rules);
+
+            expect(rule.use[1].options.silent).to.be.true;
+            expect(rule.use[1].options.happyPackMode).to.be.true;
+        });
+
+        it('configure rule for "handlebars"', () => {
+            config.enableHandlebarsLoader((options) => {
+                options.debug = true;
+            });
+            config.configureLoaderRule('handlebars', (loaderRule) => {
+                loaderRule.use[0].options.fooBar = 'fooBar';
+            });
+
+            const webpackConfig = configGenerator(config);
+            const rule = findRule(/\.(handlebars|hbs)$/, webpackConfig.module.rules);
+
+            expect(rule.use[0].options.debug).to.be.true;
+            expect(rule.use[0].options.fooBar).to.be.equal('fooBar');
         });
     });
 });